apps/backend/pyproject.toml
[project]
name = "docroot-backend"
version = "0.1.0"
description = "FastAPI backend for the documentation host"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "python-multipart>=0.0.12",
    "toml>=0.10",
    "python-jose[cryptography]>=3.3",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
